Correct import instruction
The current code is forked from the original branch but the import instruction is still the same which cause compile failure when using with vendor with Go 1.13 and GO111MODULE=off. A work around is set GO111MODULE=on.

This change propose a fix to correct the import instruction of the package to the current path of the repo.